+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 
+* Smart text extraction for retrieving layouted text from pages
 * Support for digitally signing with ECDSA keys
 * Support for digitally signing with DSA keys
 * Support for BrotliDecode filter